moved test to __test__ folder to avoid build errors
