+++ next.config.js
@@ -1,3 +1,4 @@
+/* eslint-disable no-param-reassign */
 module.exports = {
   reactStrictMode: true,
   images: {
