+++ package.json
@@ -42,6 +42,7 @@
     "jest-transform-css": "^2.1.0",
     "msw": "^0.35.0",
     "next-page-tester": "^0.28.0",
-    "node-mocks-http": "^1.10.1"
+    "node-mocks-http": "^1.10.1",
+    "webpack": "^5.19.0"
   }
 }
